fix(rome_js_formatter): remove context, move function inside proper file
